service1_main/requirements.txt
functions-framework>=3.0.0
google-cloud-storage>=2.10.0
google-cloud-logging>=3.0.0
google-cloud-pubsub>=2.18.0
